+++ changelog.md
@@ -3,3 +3,4 @@ be used, instead of the last mod in the load order having priority over which ve
 - Ability to exclude your mod from Creditor - Some developers like to split their mods into sub-plugins which are then
 all loaded as separate plugins by Hytale. You can now choose to prefix the name of your subplugins with an underscore
 to reduce clutter in the credit list.
+- Added version and asset pack tags to the plugin list.